+++ composeApp/src/androidMain/kotlin/org/jh/coach/MainActivity.kt
@@ -1,11 +1,11 @@
-package org.jh.project
+package org.jh.coach
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import com.arkivanov.decompose.ExperimentalDecomposeApi
 import com.arkivanov.decompose.retainedComponent
-import navigation.RootComponent
+import presentation.RootComponent
 
 class MainActivity : ComponentActivity() {
     @OptIn(ExperimentalDecomposeApi::class)
