+++ composeApp/src/commonMain/kotlin/presentation/firstSetupScreen/FirstSetupScreenEvent.kt
@@ -1,4 +1,4 @@
-package navigation.onboardingStack.firstSetupScreen
+package presentation.firstSetupScreen
 
 sealed interface FirstSetupScreenEvent {
     data object ClickButton : FirstSetupScreenEvent
