+++ composeApp/src/commonMain/kotlin/presentation/onboardingScreen/OnboardingScreenComponent.kt
@@ -1,4 +1,4 @@
-package navigation.onboardingStack.onboardingScreen
+package presentation.onboardingScreen
 
 import com.arkivanov.decompose.ComponentContext
