+++ composeApp/src/commonMain/kotlin/presentation/onboardingScreen/OnboardingScreenEvent.kt
@@ -1,4 +1,4 @@
-package navigation.onboardingStack.onboardingScreen
+package presentation.onboardingScreen
 
 sealed interface OnboardingScreenEvent {
     data object ClickButton : OnboardingScreenEvent
