+++ composeApp/src/commonMain/kotlin/presentation/screenB/ScreenB.kt
@@ -1,4 +1,4 @@
-package screens
+package presentation.screenB
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -8,8 +8,6 @@ import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import navigation.rootStack.screenB.ScreenBComponent
-import navigation.rootStack.screenB.ScreenBEvent
 
 @Composable
 fun ScreenB(text: String, component: ScreenBComponent) {
