+++ composeApp/src/commonMain/kotlin/presentation/screenB/ScreenBComponent.kt
@@ -1,4 +1,4 @@
-package navigation.rootStack.screenB
+package presentation.screenB
 
 import com.arkivanov.decompose.ComponentContext
