+++ composeApp/src/commonMain/kotlin/presentation/screenB/ScreenBEvent.kt
@@ -1,4 +1,4 @@
-package navigation.rootStack.screenB
+package presentation.screenB
 
 sealed interface ScreenBEvent {
     data object ClickBack : ScreenBEvent
