+++ composeApp/src/iosMain/kotlin/org/jh/coach/MainViewController.kt
@@ -1,10 +1,10 @@
-package org.jh.project
+package org.jh.coach
 
 import androidx.compose.runtime.remember
 import androidx.compose.ui.window.ComposeUIViewController
 import com.arkivanov.decompose.DefaultComponentContext
 import com.arkivanov.essenty.lifecycle.LifecycleRegistry
-import navigation.RootComponent
+import presentation.RootComponent
 
 fun MainViewController() = ComposeUIViewController {
     val root = remember {
